moved all contracts needed from the Charges solution
