+++ source/GreenEnergyHub.Charges.Libraries/source/GreenEnergyHub.DataHub.Charges.Clients/DefaultChargeLinkRequest/Contracts/CreateDefaultChargeLinkMessagesSucceeded.proto
@@ -1,0 +1,32 @@
+﻿/* Copyright 2020 Energinet DataHub A/S
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License2");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+syntax = "proto3";
+
+option csharp_namespace = "GreenEnergyHub.Charges.Commands";
+
+/*
+* This reply is sent by the Charges domain to the Metering Point domain once the request has been processed successfully.
+*
+* Expected reply meta data besides what ADR-0008 already requires for integration events:
+* https://github.com/Energinet-DataHub/green-energy-hub/blob/main/docs/architecture-decision-record/ADR-0008%20Integration%20events.md#message-metadata
+* 
+* ReplyId - UUID, generated by the sending domain
+* OriginalRequestId - For traceability
+* ProcessId - For traceability - UUID of the 'create metering point process'
+*/
+message CreateDefaultChargeLinkMessagesSucceeded {
+  string metering_point_id = 1; // The metering point that default charge link messages were created for. For the Danish version of Green Energy Hub, this is the 18-digits metering point id.
+}
