+++ source/GreenEnergyHub.Charges.Libraries/source/GreenEnergyHub.DataHub.Charges.Clients/DefaultChargeLinkRequest/Contracts/CreateDefaultChargeLinks.proto
@@ -1,0 +1,32 @@
+﻿/* Copyright 2020 Energinet DataHub A/S
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License2");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+syntax = "proto3";
+
+option csharp_namespace = "GreenEnergyHub.Charges.Commands";
+
+/*
+* This request is sent by the Metering Point domain to the Charges domain requesting default charges to be linked to a new metering point.
+*
+* Expected request meta data besides what ADR-0008 already requires for integration events:
+* https://github.com/Energinet-DataHub/green-energy-hub/blob/main/docs/architecture-decision-record/ADR-0008%20Integration%20events.md#message-metadata
+* 
+* RequestId - Generated by sending domain - UUID for the request. Will be applied to the reply message's meta data for tracing
+* ProcessId - Generated by sending domain - UUID for the 'create metering point process'. Will be applied to the reply message's meta data for tracing
+* ReplyTo - Sending domain provides a reply-path to a location where the reply must be sent
+*/
+message CreateDefaultChargeLinks {
+  string metering_point_id = 1; // Identifies the metering point to link default charges to. For the Danish version of Green Energy Hub, this is the 18-digits metering point id.
+}
