+++ source/GreenEnergyHub.Charges.Libraries/source/GreenEnergyHub.DataHub.Charges.Clients/DefaultChargeLinkRequest/Contracts/CreateDefaultChargeLinksSucceeded.proto
@@ -1,0 +1,36 @@
+﻿/* Copyright 2020 Energinet DataHub A/S
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License2");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+syntax = "proto3";
+
+option csharp_namespace = "GreenEnergyHub.Charges.Commands";
+
+/*
+* This reply is sent by the Charges domain to the Metering Point domain once the request has been processed successfully.
+*
+* Expected reply meta data besides what ADR-0008 already requires for integration events:
+* https://github.com/Energinet-DataHub/green-energy-hub/blob/main/docs/architecture-decision-record/ADR-0008%20Integration%20events.md#message-metadata
+* 
+* ReplyId - Generated by the sending domain (UUID)
+* OriginalRequestId - For traceability
+* ProcessId - For traceability - UUID of the 'create metering point process'
+*
+* A reply with did_create_charge_links = false is valid. It occurs when no default charges are defined for the given metering point type.
+* When the Metering Point domain receives above reply there is no reason to send a CreateDefaultChargeLinkMessages request.
+*/
+message CreateDefaultChargeLinksSucceeded {
+  string metering_point_id = 1; // The metering point for which default charge link were created. For the Danish version of Green Energy Hub, this is the 18-digits metering point id.
+  bool did_create_charge_links = 2; // Did the request result in any default charge links being added to the metering point (True = yes, False = no)
+}
